some fix from test '

+++ src/pages/projects/Projects.js
@@ -1,4 +1,5 @@
 import React, { useEffect } from 'react';
+import swal from 'sweetalert';
 import DashboardLayout from '../../layouts/Dasboard_Layout';
 import { useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
@@ -21,9 +22,21 @@ const Projects = () => {
   useEffect(() => {
     if (alert.message) {
       addToast(alert.message, { appearance: 'error' });
+      swal({
+        title: 'Error!',
+        text: alert.message,
+        icon: 'error',
+        button: 'close!',
+      });
     }
     if (alert.success) {
       addToast(alert.success, { appearance: 'success' });
+      swal({
+        title: 'Success!',
+        text: alert.success,
+        icon: 'success',
+        button: 'close!',
+      });
     }
   }, [alert.message, alert.success, addToast]);
 
@@ -34,7 +47,20 @@ const Projects = () => {
   };
 
   const deleteAProject = (id) => {
-    dispatch(deleteProject(id));
+    swal({
+      title: 'Are you sure?',
+      text: 'Once deleted, you will not be able to recover this file!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          dispatch(deleteProject(id));
+        } else {
+          swal('Operation canceled!');
+        }
+      });
   };
 
   return (
